# >2x faster cholesky inverse and avoid oom when possible